✨feat: 상품 엔티티에 평균 평점 컬럼 추가 및 평균 평점 업데이트 기능 추가

+++ src/main/java/com/project/openmarket/domain/review/repository/ReviewRepository.java
@@ -16,5 +16,5 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
 	List<Review> findByConsumer(Consumer consumer);
 
 	@Query(value = "select avg(r.score) from review r where r.product = :product group by r.product", nativeQuery = true)
-	long getAvgScoreByProduct(@Param("product") Product product);
+	double getAvgScoreByProduct(@Param("product") Product product);
 }
